requirements.txt
Flask
flask_bootstrap
gunicorn
pylint
google-cloud-translate==1.7.0